vmm: make possible to set page table options with flags
With flags, you can set that page table options, e.g, read only or
cacheable. The flags are modeled after the x86 page options, but might
be portable (I want them to be :P )

Also added a function for unmapping page, but it only calls the private
function that did the same thing.