smbios: map a page only if it was in memory map
Generally, the 0xf0000+ range are in the memory map, while others
aren't. This small fix only will map a memory area for smbios if it
doesn't exist in the memory map

WARNING: This is a fix. It might have better ways to do this (like
turning able to you specify the address yourself for those kind of
operations, bypassing the memory manager, or making the MM consider
those regions (by getting PCI memory maps?)